requirements.txt
certifi==2020.6.20
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
gunicorn==20.1.0
numpy == 1.19.3
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
Werkzeug==2.0.1
tensorflow-cpu==2.5.0